Operational transforms for adding and removing lists
This handles transforms between Add/Remove list and other operations that
modify the number of steps in the document. There are some unresolved
conflicts with merging paragraphs and adding lists that return null as the
transform result which required merge and split operations on lists to be
resolved correctly.

Added tests for the new parts of the op transform matrix and changed the
op transform test code to parse any op parameter containing "length" or
"position" in the name as an integer.